requirements.txt
streamlit
pandas
plotly
google-genai
requests